requirements.txt
streamlit
pandas
openai>=1.0.0
google-generativeai
pre-commit
pytest-mock
pytest